criando context pra atualizar tela meus dados quando algum dado for alterado
