+++ src/core/dto/dados-usuario-dto.ts
@@ -1,9 +1,9 @@
 export type DadosUsuarioDTO = {
-  nome: string;
-  cpf: string;
-  login: string;
-  email: string;
-  nomeUnidade: string;
-  emailEducacional: string;
-  tipo: number;
+  nome?: string;
+  cpf?: string;
+  login?: string;
+  email?: string;
+  nomeUnidade?: string;
+  emailEducacional?: string;
+  tipo?: number;
 };
