+++ src/routes/index.tsx
@@ -24,7 +24,7 @@ import Home from '~/pages/home';
 import Inicial from '~/pages/inicial';
 import Login from '~/pages/login';
 import LoginAutomatico from '~/pages/login-automatico';
-import MeusDados from '~/pages/meus-dados';
+import { MeusDadosProvider } from '~/pages/meus-dados';
 import Principal from '~/pages/principal/index';
 import RedefinirSenha from '~/pages/redefinir-senha';
 import RedefinirSenhaToken from '~/pages/redefinir-senha-token';
@@ -40,7 +40,6 @@ const RoutesConfig = () => {
   const principalPage = createElement(Principal);
   const forbiddenPage = createElement(PageForbidden);
   const iniciallPage = createElement(Inicial);
-  const meusDadosPage = createElement(MeusDados);
   const redefinirSenhaPage = createElement(RedefinirSenha);
   const redefinirSenhaTokenPage = createElement(RedefinirSenhaToken);
 
@@ -77,7 +76,7 @@ const RoutesConfig = () => {
                 <Route path={ROUTES.SEM_PERMISSAO} element={forbiddenPage} />
                 <Route path={ROUTES.PRINCIPAL} element={iniciallPage} />
                 <Route path={ROUTES.LOGIN} element={<Navigate to={ROUTES.PRINCIPAL} />} />
-                <Route path={ROUTES.MEUS_DADOS} element={meusDadosPage} />
+                <Route path={ROUTES.MEUS_DADOS} element={<MeusDadosProvider />} />
                 <Route path={ROUTES.INSCRICAO} element={<Inscricao />} />
 
                 <Route path={ROUTES.AREA_PROMOTORA}>
